working on complete order, not finished, just a push in case anyone wants to work on it while I'm gone.
