+++ website/templates/index.html
@@ -18,7 +18,6 @@
   <li role="presentation"><a href="{% url 'website:sell' %}">Sell a Product</a></li>
   <li role="presentation"><a href="{% url 'website:shopping_cart' %}">Shopping Cart</a></li>
   <li role="presentation"><a href="{% url 'website:payment_type' %}">Add A Payment Type</a></li>
-  <li role="presentation"><a href="{% url 'website:paymentlist' %}">Payment List</a></li>
   <li role="presentation"><a href="{% url 'website:logout' %}">Logout</a></li>
 </ul>
 {% else %}
