+++ website/templates/payment.html
@@ -8,6 +8,6 @@
 	      {% csrf_token %}
 	      {{ payment_type_form.as_p }}
 	      <input type="submit" value="Submit" />
-	    </form>
+	  </form>
         {% endif %}
 {% endblock %}
