+++ website/views/views.py
@@ -148,14 +148,4 @@ def add_payment_type(request):
             account_number=form_data['account_number'],
         )
         p.save()
-        template_name = 'payment.html'
-        return render(request, template_name, {'paymenttype': form_data})
-
-def all_payment_types(request):
-        user = request.user
-        all_payment_types = PaymentType.objects.filter(user_id=user.id)
-        template_name = 'list_payment.html'
-        payment_type_dict = {'all_payment_types': all_payment_types}
-        return render(request, template_name, payment_type_dict)
-
- 
+        return HttpResponseRedirect('/categories')
